限制数字配置取值范围

+++ backend/src/main/java/com/agentnexus/backend/tag/api/dto/request/TagInput.java
@@ -1,14 +1,15 @@
 package com.agentnexus.backend.tag.api.dto.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record TagInput(
     String tagName,
     String tagType,
     String description,
-    Integer minValue,
-    Integer maxValue,
-    Integer passThreshold,
+    BigDecimal minValue,
+    BigDecimal maxValue,
+    BigDecimal passThreshold,
     List<TagOptionInput> options
 ) {
 }
